feat: merge sanitizers
