--- libs/@local/hashql/ast/tests/ui/lowering/generic-constraint-sanitizer/.spec.toml
@@ -1,1 +1,0 @@
-suite = "ast/lowering/generic-constraint-sanitizer"
