+++ libs/@local/hashql/ast/tests/ui/lowering/sanitizer/.spec.toml
@@ -1,0 +1,1 @@
+suite = "ast/lowering/sanitizer"
